+++ stac_fastapi/tests/api/test_api.py
@@ -1,6 +1,5 @@
-import copy
 import uuid
-from datetime import datetime, timedelta
+from datetime import timedelta
 
 import pytest
 
@@ -28,7 +27,7 @@ ROUTES = {
     "DELETE /collections/{collection_id}/items/{item_id}",
     "POST /collections",
     "POST /collections/{collection_id}/items",
-    "PUT /collections",
+    "PUT /collections/{collection_id}",
     "PUT /collections/{collection_id}/items/{item_id}",
 }
 
@@ -62,11 +61,11 @@ async def test_router(app):
 
 
 @pytest.mark.asyncio
-async def test_app_transaction_extension(app_client, ctx):
-    item = copy.deepcopy(ctx.item)
+async def test_app_transaction_extension(app_client, ctx, load_test_data):
+    item = load_test_data("test_item.json")
     item["id"] = str(uuid.uuid4())
     resp = await app_client.post(f"/collections/{item['collection']}/items", json=item)
-    assert resp.status_code == 200
+    assert resp.status_code == 201
 
     await app_client.delete(f"/collections/{item['collection']}/items/{item['id']}")
 
@@ -84,11 +83,11 @@ async def test_app_search_response(app_client, ctx):
 
 
 @pytest.mark.asyncio
-async def test_app_context_extension(app_client, ctx, txn_client):
-    test_item = ctx.item
+async def test_app_context_extension(app_client, txn_client, ctx, load_test_data):
+    test_item = load_test_data("test_item.json")
     test_item["id"] = "test-item-2"
     test_item["collection"] = "test-collection-2"
-    test_collection = ctx.collection
+    test_collection = load_test_data("test_collection.json")
     test_collection["id"] = "test-collection-2"
 
     await create_collection(txn_client, test_collection)
@@ -109,6 +108,7 @@ async def test_app_context_extension(app_client, ctx, txn_client):
 
     resp = await app_client.post("/search", json={"collections": ["test-collection-2"]})
     assert resp.status_code == 200
+
     resp_json = resp.json()
     assert len(resp_json["features"]) == 1
     assert "context" in resp_json
@@ -127,10 +127,11 @@ async def test_app_fields_extension(app_client, ctx, txn_client):
 
 @pytest.mark.asyncio
 async def test_app_fields_extension_query(app_client, ctx, txn_client):
+    item = ctx.item
     resp = await app_client.post(
         "/search",
         json={
-            "query": {"proj:epsg": {"gte": ctx.item["properties"]["proj:epsg"]}},
+            "query": {"proj:epsg": {"gte": item["properties"]["proj:epsg"]}},
             "collections": ["test-collection"],
         },
     )
@@ -181,8 +182,10 @@ async def test_app_fields_extension_no_null_fields(app_client, ctx, txn_client):
 
 
 @pytest.mark.asyncio
-async def test_app_fields_extension_return_all_properties(app_client, ctx, txn_client):
-    item = ctx.item
+async def test_app_fields_extension_return_all_properties(
+    app_client, ctx, txn_client, load_test_data
+):
+    item = load_test_data("test_item.json")
     resp = await app_client.get(
         "/search", params={"collections": ["test-collection"], "fields": "properties"}
     )
@@ -223,7 +226,9 @@ async def test_app_query_extension_gte(app_client, ctx):
 
 @pytest.mark.asyncio
 async def test_app_query_extension_limit_lt0(app_client):
-    assert (await app_client.post("/search", json={"limit": -1})).status_code == 400
+    assert (
+        await app_client.post("/search", json={"query": {}, "limit": -1})
+    ).status_code == 400
 
 
 @pytest.mark.asyncio
@@ -243,16 +248,14 @@ async def test_app_query_extension_limit_10000(app_client):
 @pytest.mark.asyncio
 async def test_app_sort_extension_get_asc(app_client, txn_client, ctx):
     first_item = ctx.item
-    item_date = datetime.strptime(
-        first_item["properties"]["datetime"], "%Y-%m-%dT%H:%M:%SZ"
-    )
 
     second_item = dict(first_item)
     second_item["id"] = "another-item"
-    another_item_date = item_date - timedelta(days=1)
-    second_item["properties"]["datetime"] = another_item_date.strftime(
-        "%Y-%m-%dT%H:%M:%SZ"
+    another_item_date = first_item["properties"]["datetime"] - timedelta(days=1)
+    second_item["properties"]["datetime"] = another_item_date.isoformat().replace(
+        "+00:00", "Z"
     )
+
     await create_item(txn_client, second_item)
 
     resp = await app_client.get("/search?sortby=+properties.datetime")
@@ -265,15 +268,12 @@ async def test_app_sort_extension_get_asc(app_client, txn_client, ctx):
 @pytest.mark.asyncio
 async def test_app_sort_extension_get_desc(app_client, txn_client, ctx):
     first_item = ctx.item
-    item_date = datetime.strptime(
-        first_item["properties"]["datetime"], "%Y-%m-%dT%H:%M:%SZ"
-    )
 
     second_item = dict(first_item)
     second_item["id"] = "another-item"
-    another_item_date = item_date - timedelta(days=1)
-    second_item["properties"]["datetime"] = another_item_date.strftime(
-        "%Y-%m-%dT%H:%M:%SZ"
+    another_item_date = first_item["properties"]["datetime"] - timedelta(days=1)
+    second_item["properties"]["datetime"] = another_item_date.isoformat().replace(
+        "+00:00", "Z"
     )
     await create_item(txn_client, second_item)
 
@@ -287,15 +287,12 @@ async def test_app_sort_extension_get_desc(app_client, txn_client, ctx):
 @pytest.mark.asyncio
 async def test_app_sort_extension_post_asc(app_client, txn_client, ctx):
     first_item = ctx.item
-    item_date = datetime.strptime(
-        first_item["properties"]["datetime"], "%Y-%m-%dT%H:%M:%SZ"
-    )
 
     second_item = dict(first_item)
     second_item["id"] = "another-item"
-    another_item_date = item_date - timedelta(days=1)
-    second_item["properties"]["datetime"] = another_item_date.strftime(
-        "%Y-%m-%dT%H:%M:%SZ"
+    another_item_date = first_item["properties"]["datetime"] - timedelta(days=1)
+    second_item["properties"]["datetime"] = another_item_date.isoformat().replace(
+        "+00:00", "Z"
     )
     await create_item(txn_client, second_item)
 
@@ -313,15 +310,12 @@ async def test_app_sort_extension_post_asc(app_client, txn_client, ctx):
 @pytest.mark.asyncio
 async def test_app_sort_extension_post_desc(app_client, txn_client, ctx):
     first_item = ctx.item
-    item_date = datetime.strptime(
-        first_item["properties"]["datetime"], "%Y-%m-%dT%H:%M:%SZ"
-    )
 
     second_item = dict(first_item)
     second_item["id"] = "another-item"
-    another_item_date = item_date - timedelta(days=1)
-    second_item["properties"]["datetime"] = another_item_date.strftime(
-        "%Y-%m-%dT%H:%M:%SZ"
+    another_item_date = first_item["properties"]["datetime"] - timedelta(days=1)
+    second_item["properties"]["datetime"] = another_item_date.isoformat().replace(
+        "+00:00", "Z"
     )
     await create_item(txn_client, second_item)