+++ stac_fastapi/tests/extensions/test_filter.py
@@ -34,7 +34,9 @@ async def test_search_filter_extension_eq_get(app_client, ctx):
 
 @pytest.mark.asyncio
 async def test_search_filter_extension_eq_post(app_client, ctx):
-    params = {"filter": {"op": "=", "args": [{"property": "id"}, ctx.item["id"]]}}
+    params = {
+        "filter": {"op": "=", "args": [{"property": "id"}, ctx.item["id"]]},
+    }
     resp = await app_client.post("/search", json=params)
     assert resp.status_code == 200
     resp_json = resp.json()